Migrate ZipExtractor logic to AutoUpdater library
