+++ AutoUpdater.NET/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="ZipStorer" version="3.4.0" targetFramework="net40" />
+</packages>
